requirements.txt
# Core Deep Learning
torch==2.1.0
torchvision==0.16.0
transformers==4.36.0
accelerate==0.25.0
bitsandbytes==0.41.3

# Vision & OCR
qwen-vl-utils==0.0.3
Pillow==10.1.0
opencv-python==4.8.1.78

# YOLO Detection
ultralytics==8.0.228

# PDF Processing
pypdfium2==4.25.0

# Utilities
numpy==1.24.3
pandas==2.1.3
matplotlib==3.8.2
tqdm==4.66.1
seaborn==0.13.0          ← THIS LINE NEEDED FOR EDA!

# Optional
streamlit==1.29.0
